Split config tests into a test for internal external functions
